latest_ai_development/pyproject.toml
[project]
name = "latest_ai_development"
version = "0.1.0"
description = "latest-ai-development using crewAI"
authors = [{ name = "Frank Rumler", email = "frank@rumler.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.130.0,<1.0.0",
    "PyYAML>=6.0.2,<7.0.0",
    "python-dotenv>=1.0.1,<2.0.0"
]

[project.scripts]
latest_ai_development = "latest_ai_development.main:run"
run_crew = "latest_ai_development.main:run"
train = "latest_ai_development.main:train"
replay = "latest_ai_development.main:replay"
test = "latest_ai_development.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
